+++ customHooks/useRipples/callback/fade.ts
@@ -14,25 +14,19 @@ export function fade(this: Ripples) {
   const { toBeList } = fadeData;
 
   if (isNull(renderData)) {
-    fadeData.isTransitioning = false;
-    dog.warn('还没有准备好，即将退出');
-    Reflect.apply(runSide, this, []);
-    return;
+    return Reflect.apply(exitFade, this, ['还没有准备好，即将退出']);
   }
+
   if (isEmptyArray(toBeList)) {
-    fadeData.isTransitioning = false;
-    dog.error('由于缺少当前渲染的背景，退出');
-    Reflect.apply(runSide, this, []);
-    return;
+    return Reflect.apply(exitFade, this, ['由于缺少当前渲染的背景，退出']);
   }
+
   if (fadeData.drawProgress === 0) {
     dog('开始执行渐变，当前尚有可执行', toBeList.length);
     toBeList.forEach(e => dog(e));
   }
   // 进度完成则结束当前的进度
   if (fadeData.drawProgress > 1000) {
-    // 临时关闭当前的状体
-    fadeData.isTransitioning = false;
     // 最后使用的纹理
     fadeData.lastDrawImage = toBeList.shift()!;
     dog('执行渐变背景完毕，剩余可执行', toBeList);
@@ -42,7 +36,6 @@ export function fade(this: Ripples) {
     // 尚有未执行完毕的
     if (isEmptyArray(toBeList)) {
       const { lastUseStyle } = renderData;
-
       // 没有设置背景色或是背景图
       if (
         (isNull(options.imgUrl) ||
@@ -53,12 +46,10 @@ export function fade(this: Ripples) {
       ) {
         // 因为此时渲染为空，需要手动添加一个默认渲染
         Reflect.apply(setTransparentTexture, this, [false]);
-        Reflect.apply(runSide, this, []);
       }
-    } else
-      // 启用下一轮的循环
-      Reflect.apply(runSide, this, []);
-    return;
+    }
+    // 启用下一轮的循环
+    return Reflect.apply(exitFade, this, ['当前执行完毕，开启下次执行']);
   }
   const { lastDrawImage, backgroundInfo } = fadeData;
 
@@ -67,8 +58,16 @@ export function fade(this: Ripples) {
 
   const canvas = document.createElement('canvas');
   const ctx = canvas.getContext('2d');
+  /**
+   *  bug: 2508021258（2）
+   *
+   *  这里判定了当前的执行环境并退出了执行，但是并没有设定 `isTransitioning` 值
+   *
+   */
   if (isNull(ctx) || isNull(lastDrawImage.resource) || isEmptyArray(toBeList)) {
+    fadeData.isTransitioning = false;
     dog.error('渲染出现意外，该有的值不存在导致 bug');
+    Reflect.apply(runSide, this, []);
     return;
   }
   const { width, height } = backgroundInfo;
@@ -78,8 +77,21 @@ export function fade(this: Ripples) {
   // ctx.globalCompositeOperation = 'destination-over';
   {
     // ctx.globalAlpha = 1 - drawProgress;
-    ctx.drawImage(lastDrawImage.resource, 0, 0, width, height);
-    ctx.globalAlpha = fadeData.drawProgress / 1000;
+    ctx.drawImage(lastDrawImage.resource, 0, 0, width, height); // 绘制上一次的图案
+    ctx.globalAlpha = fadeData.drawProgress / 1000; // 设置透明度
+    /**
+     * bug: 2508021258（1）
+     *
+     * TODO 这里 报错
+     *
+     * Cannot read properties of undefined (reading 'resource')
+     *
+     *
+     * 报错说这里没有数据，在复现的 bug 数据中，数组第一位为 undefined，而 2、3 位有真实数据
+     *
+     * 确认该错误是由于在判定 `fadeData.isTransitioning` 为 `true` 造成的第一个待执行项不存在却被设置为新的数组的第一项
+     *
+     */
     ctx.drawImage(toBeList[0].resource, 0, 0, width, height);
   }
 
@@ -101,4 +113,12 @@ export function fade(this: Ripples) {
     fadeData.drawProgress = 0;
   }
   dog.type = true;
+}
+
+/**  退出当前的渐变执行  */
+function exitFade(this: Ripples, message: string) {
+  const { fadeData } = this;
+  fadeData.isTransitioning = false;
+  dog.warn(message);
+  Reflect.apply(runSide, this, []);
 }