+++ scripts/pub.sh
@@ -1,9 +1,20 @@
 #!/bin/bash
 
 CHECK_VERSION="@qqi/check-version"
-printf $(pnpm dlx "${CHECK_VERSION}" -v)  # 更改全局安装的测试方法
+# 安装  
+install_check_version() {
+    if ! npm  list -g --depth=0 | grep -q "${CHECK_VERSION}"; then 
+        echo "当前未全局安装 '${CHECK_VERSION}'，即将进行安装"
+        npm install ${CHECK_VERSION} --global
+    else 
+         echo "包 ${CHECK_VERSION} 已全局安装"
+    fi
+}
+
 tag=""
-if ! tag=$(pnpm dlx "${CHECK_VERSION}" c=. 2>&1); then
+# npx 在某种程度上还是要比 pnpm 更靠谱呀
+install_check_version
+if ! tag=$(npx "${CHECK_VERSION}" c=. 2>&1); then
     echo "未通过版本校验：$tag"
     exit 0
 fi